refactor: remove useless subpackages in tooltip.impl
