+++ common/src/main/java/com/wynntils/handlers/tooltip/TooltipHandler.java
@@ -5,9 +5,9 @@
 package com.wynntils.handlers.tooltip;
 
 import com.wynntils.core.components.Handler;
-import com.wynntils.handlers.tooltip.impl.charm.CharmTooltipComponent;
-import com.wynntils.handlers.tooltip.impl.gear.GearTooltipComponent;
-import com.wynntils.handlers.tooltip.impl.tome.TomeTooltipComponent;
+import com.wynntils.handlers.tooltip.impl.CharmTooltipComponent;
+import com.wynntils.handlers.tooltip.impl.GearTooltipComponent;
+import com.wynntils.handlers.tooltip.impl.TomeTooltipComponent;
 import com.wynntils.handlers.tooltip.type.TooltipComponent;
 import com.wynntils.models.items.WynnItem;
 import com.wynntils.models.items.items.game.CharmItem;
