+++ common/src/main/java/com/wynntils/handlers/tooltip/impl/GearTooltipComponent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.handlers.tooltip.impl.gear;
+package com.wynntils.handlers.tooltip.impl;
 
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
